requirements.txt
allure-pytest
argparse
configcat-client>=5.0.0
configparser~=5.0.0
contextvars
Django~=3.1.5
pactman>=2.28.0
pdfminer
pip
PyPDF2
python-dateutil~=2.8.1
pytz
requests~=2.25.1
setuptools
urllib3
